Change the way we do some counting..In some cases the block heap knows enough about the counts that we don't need to track them otherwise.  Also change some variables to use size_t instead of unsigned long
git-svn-id: http://svn.ratbox.org/svnroot/trunk@13319 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ include/client.h
@@ -564,8 +564,8 @@ extern int            exit_client(struct Client*, struct Client*,
                                   struct Client*, const char*);
 
 
-extern void     count_local_client_memory(int *count, int *memory);
-extern void     count_remote_client_memory(int *count, int *memory);
+extern void     count_local_client_memory(size_t *count, size_t *memory);
+extern void     count_remote_client_memory(size_t *count, size_t *memory);
 
 extern struct Client* find_chasing (struct Client *, char *, int *);
 extern struct Client* find_person (char *);
